Bug 54011 - LargeFileAccess: use POSIX on OS X
The old Mac LargeFileAccess code was using a set of deprecated
APIs, and also had a tendency to fail for no discernible reason.
Since the Mac-specific code appears to have been a workaround for
a bug in CodeWarrior anyway, use the now-perfectly functional
POSIX API instead.

The only new function provided as a part of this patch is a port
of the resource fork logic to the POSIX API.

Signed-off-by: Hubert Figuière <hub@figuiere.net>