Bug 58175 - Replace OS X FlatCarbon headers
Several files were including FlatCarbon headers with a hardcoded
include path. Unfortunately, this path is no longer valid as of recent
Xcode versions.

Fortunately, the FlatCarbon headers are just shims for the real
includes, and all of the includes exempi uses come from CoreServices.
Replacing the specific includes with the generic CoreServices header
is a more compatible solution.

Signed-off-by: Hubert Figuière <hub@figuiere.net>