Drop old/obsolete symfile
This would, at least on some setups, result in actual API symbols
missing from the library, because the symfile hasn't been changed
for over a decade.

Additionally, it should not be needed anymore as the project
appears to use `-fvisibility=hidden` and actually tag the public
API symbols appropriately.

This fixes tests build with LTO (as long as static libs are not
disabled) besides other issues.